+++ assemblyadapter-recycler/src/androidTest/java/com/github/panpf/assemblyadapter/recycler/test/divider/AssemblyStaggeredGridDividerItemDecorationTest.kt
@@ -19,15 +19,94 @@ import android.graphics.Color
 import android.graphics.drawable.ColorDrawable
 import androidx.recyclerview.widget.RecyclerView
 import androidx.test.platform.app.InstrumentationRegistry
-import com.github.panpf.assemblyadapter.recycler.divider.AssemblyDividerConfig
-import com.github.panpf.assemblyadapter.recycler.divider.AssemblyStaggeredGridDividerItemDecoration
-import com.github.panpf.assemblyadapter.recycler.divider.Divider
+import com.github.panpf.assemblyadapter.ItemFactory
+import com.github.panpf.assemblyadapter.recycler.divider.*
 import com.github.panpf.tools4j.test.ktx.assertNoThrow
 import com.github.panpf.tools4j.test.ktx.assertThrow
 import org.junit.Assert
 import org.junit.Test
 
 class AssemblyStaggeredGridDividerItemDecorationTest {
+
+    @Test
+    fun testConstructor(){
+        val context = InstrumentationRegistry.getInstrumentation().context
+
+        /*
+         * sideDivider personalise size exception
+         */
+        AssemblyGridDividerItemDecoration.Builder(context).apply {
+            sideDivider(Divider.space(5, Insets.allOf(2))) {
+                personaliseByItemFactoryClass(ItemFactory::class, Divider.space(5, Insets.allOf(2)))
+            }
+        }.build()
+        assertThrow(IllegalArgumentException::class) {
+            AssemblyGridDividerItemDecoration.Builder(context).apply {
+                sideDivider(Divider.space(5, Insets.allOf(2))) {
+                    personaliseByItemFactoryClass(ItemFactory::class, Divider.space(4, Insets.allOf(2)))
+                }
+            }.build()
+        }
+        assertThrow(IllegalArgumentException::class) {
+            AssemblyGridDividerItemDecoration.Builder(context).apply {
+                sideDivider(Divider.space(5, Insets.allOf(2))) {
+                    personaliseByItemFactoryClass(ItemFactory::class, Divider.space(5, Insets.allOf(3)))
+                }
+            }.build()
+        }
+
+        /*
+         * sideHeaderDivider personalise size exception
+         */
+        AssemblyGridDividerItemDecoration.Builder(context).apply {
+            sideDivider(Divider.space(5, Insets.allOf(2)))
+            sideHeaderDivider(Divider.space(5, Insets.allOf(2))) {
+                personaliseByItemFactoryClass(ItemFactory::class, Divider.space(5, Insets.allOf(2)))
+            }
+        }.build()
+        assertThrow(IllegalArgumentException::class) {
+            AssemblyGridDividerItemDecoration.Builder(context).apply {
+                sideDivider(Divider.space(5, Insets.allOf(2)))
+                sideHeaderDivider(Divider.space(5, Insets.allOf(2))) {
+                    personaliseByItemFactoryClass(ItemFactory::class, Divider.space(4, Insets.allOf(2)))
+                }
+            }.build()
+        }
+        assertThrow(IllegalArgumentException::class) {
+            AssemblyGridDividerItemDecoration.Builder(context).apply {
+                sideDivider(Divider.space(5, Insets.allOf(2)))
+                sideHeaderDivider(Divider.space(5, Insets.allOf(2))) {
+                    personaliseByItemFactoryClass(ItemFactory::class, Divider.space(5, Insets.allOf(3)))
+                }
+            }.build()
+        }
+
+        /*
+         * sideFooterDivider personalise size exception
+         */
+        AssemblyGridDividerItemDecoration.Builder(context).apply {
+            sideDivider(Divider.space(5, Insets.allOf(2)))
+            sideFooterDivider(Divider.space(5, Insets.allOf(2))) {
+                personaliseByItemFactoryClass(ItemFactory::class, Divider.space(5, Insets.allOf(2)))
+            }
+        }.build()
+        assertThrow(IllegalArgumentException::class) {
+            AssemblyGridDividerItemDecoration.Builder(context).apply {
+                sideDivider(Divider.space(5, Insets.allOf(2)))
+                sideFooterDivider(Divider.space(5, Insets.allOf(2))) {
+                    personaliseByItemFactoryClass(ItemFactory::class, Divider.space(4, Insets.allOf(2)))
+                }
+            }.build()
+        }
+        assertThrow(IllegalArgumentException::class) {
+            AssemblyGridDividerItemDecoration.Builder(context).apply {
+                sideDivider(Divider.space(5, Insets.allOf(2)))
+                sideFooterDivider(Divider.space(5, Insets.allOf(2))) {
+                    personaliseByItemFactoryClass(ItemFactory::class, Divider.space(5, Insets.allOf(3)))
+                }
+            }.build()
+        }
+    }
 
     @Test
     fun testIsFullSpanByPosition() {