+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * change: Now \*DividerItemDecoration useDividerAs\*Divider is true but an exception will be thrown when the corresponding divider is null
 * change: Now GridDividerItemDecoration and StaggeredGridDividerItemDecoration will throw an exception when there is sideHeaderDivider or sideFooter but no sideDivider
 * change: Now GridDividerItemDecoration and StaggeredGridDividerItemDecoration no longer allow sideHeaderDivider or sideFooterDivider and sideDivider to have different sizes and insets
+* change: Now the personalized configuration of sideDivider, sideHeaderDivider, and sideFooterDivider of GridDividerItemDecoration and StaggeredGridDividerItemDecoration no longer allows different sizes and insets
 
 # v4.0.1